+++ apps/test/.eslintrc.js
@@ -1,5 +1,5 @@
 module.exports = {
-    ...require('@squares/configs/eslint/eslint.json'),
+    ...require('@squaress/configs/eslint/eslint.json'),
     root: true,
     extends: ['plugin:tailwindcss/recommended', 'next/core-web-vitals'],
 }