+++ apps/test/tailwind.config.ts
@@ -1,4 +1,4 @@
-import { theme } from '@squares/ui/theme'
+import { theme } from '@squaress/ui/theme'
 import type { Config } from 'tailwindcss'
 
 const config: Config = {
@@ -6,7 +6,7 @@ const config: Config = {
         './src/pages/**/*.{js,ts,jsx,tsx,mdx}',
         './src/components/**/*.{js,ts,jsx,tsx,mdx}',
         './src/app/**/*.{js,ts,jsx,tsx,mdx}',
-        './node_modules/@squares/ui/**/*.{js,jsx,ts,tsx}',
+        './node_modules/@squaress/ui/**/*.{js,jsx,ts,tsx}',
     ],
     theme: theme,
     plugins: [],